+++ src/revisionGraph/flow/index.ts
@@ -12,6 +12,7 @@ export * from './flowEqualizationTarget';
 export * from './flowPullRequestContext';
 export * from './flowPullRequestPreflight';
 export * from './flowPullRequestTargets';
+export * from './flowPatternSafety';
 export * from './flowReleaseBranch';
 export * from './flowBranchStartPreflight';
 export * from './flowState';